move report related to its own module
